Separate POST secrets into two routes
