+++ common/express/sendJSONResponse.js
@@ -3,6 +3,7 @@ var self = sendJSONResponse;
 module.exports = self;
 global.sendJSONResponse = self;
 
-function sendJSONResponse(res, obj) {
-  return res.status(200).json(obj);
+function sendJSONResponse(res, obj, statusCode) {
+  var responseCode = statusCode || 200;
+  return res.status(responseCode).json(obj);
 }
